requirements.txt
# Bayes
pymc==5.10.3
ipython==8.20.0
ipykernel==6.28.0
jupyterlab==4.0.10
matplotlib==3.8.2
redis==5.0.1
python-decouple==3.8
